Modificando o APP, inserindo o schemas, models.
