+++ backend/main.py
@@ -1,15 +1,9 @@
-from itertools import product
-from fastapi import FastAPI, status
-from fastapi import HTTPException
-from database import Base, engine, Product
-from pydantic import BaseModel
+from fastapi import FastAPI, status, HTTPException
+from database import Base, engine
 from sqlalchemy.orm import Session
+import models
+import schemas
 
-# Create ToDoRequest Base Model
-class ProductRequest(BaseModel):
-    name: str
-    price: int
-    serie: int
 
 
 # Create the database
@@ -25,12 +19,12 @@ def root():
     return "produto"
 
 @app.post("/product", status_code=status.HTTP_201_CREATED)
-def create_product(product: ProductRequest):
+def create_product(product: schemas.Product):
  # create a new database session
     session = Session(bind=engine, expire_on_commit=False)
 
     # create an instance of the ToDo database model
-    productdb = Product(name = product.name, price = product.price, serie = product.serie)
+    productdb = models.Product(name = product.name, price = product.price, serie = product.serie)
 
     # add it to the session and commit it
     session.add(productdb)
@@ -51,7 +45,7 @@ def read_product(id: int):
     session = Session(bind=engine, expire_on_commit=False)
 
     # get the todo item with the given id
-    product = session.query(Product).get(id)
+    product = session.query(models.Product).get(id)
 
     # check if todo item with given id exists. If not, raise exception and return 404 not found response
     if not product:
@@ -66,7 +60,7 @@ def update_product(id: int, name: str, price:int, serie: int):
     session = Session(bind=engine, expire_on_commit=False)
 
     # get the todo item with the given id
-    product = session.query(Product).get(id)
+    product = session.query(models.Product).get(id)
 
     # update todo item with the given task (if an item with the given id was found)
     if product:
@@ -92,7 +86,7 @@ def delete_product(id: int):
     session = Session(bind=engine, expire_on_commit=False)
 
     # get the todo item with the given id
-    product = session.query(Product).get(id)
+    product = session.query(models.Product).get(id)
 
     # if todo item with given id exists, delete it from the database. Otherwise raise 404 error
     if product:
@@ -110,7 +104,7 @@ def read_product_list():
     session = Session(bind=engine, expire_on_commit=False)
 
     # get all todo items
-    product_list = session.query(Product).all()
+    product_list = session.query(models.Product).all()
 
     # close the session
     session.close()
