+++ backend/schemas.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+# Create ToDoRequest Base Model
+class Product(BaseModel):
+    name: str
+    price: int
+    serie: int
